+++ src/main/java/io/supertokens/storage/mysql/queries/EmailPasswordQueries.java
@@ -41,6 +41,12 @@ public class EmailPasswordQueries {
                 + "is_email_verified TINYINT(1) NOT NULL," + "PRIMARY KEY (user_id));";
     }
 
+    static String getQueryToCreateUserPaginationIndex(Start start) {
+        return "CREATE INDEX emailpassword_user_pagination_index ON " +
+                Config.getConfig(start).getUsersTable() + "(time_joined DESC, user_id " +
+                "DESC);";
+    }
+
     static String getQueryToCreatePasswordResetTokensTable(Start start) {
         return "CREATE TABLE IF NOT EXISTS " + Config.getConfig(start).getPasswordResetTokensTable() + " ("
                 + "user_id CHAR(36) NOT NULL," + "token VARCHAR(128) NOT NULL UNIQUE,"
@@ -354,6 +360,65 @@ public class EmailPasswordQueries {
             }
         }
         return null;
+    }
+
+    public static UserInfo[] getUsersInfo(Start start, Integer limit, String timeJoinedOrder)
+            throws SQLException, StorageQueryException {
+        String QUERY = "SELECT user_id, email, password_hash, time_joined, is_email_verified FROM "
+                + Config.getConfig(start).getUsersTable() + " ORDER BY time_joined " + timeJoinedOrder + ", user_id DESC LIMIT ?";
+        try (Connection con = ConnectionPool.getConnection(start);
+             PreparedStatement pst = con.prepareStatement(QUERY)) {
+            pst.setInt(1, limit);
+            ResultSet result = pst.executeQuery();
+            List<UserInfo> temp = new ArrayList<>();
+            while (result.next()) {
+                temp.add(UserInfoRowMapper.getInstance().mapOrThrow(result));
+            }
+            UserInfo[] finalResult = new UserInfo[temp.size()];
+            for (int i = 0; i < temp.size(); i++) {
+                finalResult[i] = temp.get(i);
+            }
+            return finalResult;
+        }
+    }
+
+    public static UserInfo[] getUsersInfo(Start start, String userId, Long timeJoined, Integer limit, String timeJoinedOrder)
+            throws SQLException, StorageQueryException {
+        String timeJoinedOrderSymbol = timeJoinedOrder.equals("ASC") ? ">" : "<";
+        String QUERY = "SELECT user_id, email, password_hash, time_joined, is_email_verified FROM "
+                + Config.getConfig(start).getUsersTable() +
+                " WHERE time_joined " + timeJoinedOrderSymbol + " ? OR (time_joined = ? AND user_id <= ?) ORDER BY time_joined " + timeJoinedOrder + ", user_id DESC LIMIT ?";
+        try (Connection con = ConnectionPool.getConnection(start);
+             PreparedStatement pst = con.prepareStatement(QUERY)) {
+            pst.setLong(1, timeJoined);
+            pst.setLong(2, timeJoined);
+            pst.setString(3, userId);
+            pst.setInt(4, limit);
+            ResultSet result = pst.executeQuery();
+            List<UserInfo> temp = new ArrayList<>();
+            while (result.next()) {
+                temp.add(UserInfoRowMapper.getInstance().mapOrThrow(result));
+            }
+            UserInfo[] finalResult = new UserInfo[temp.size()];
+            for (int i = 0; i < temp.size(); i++) {
+                finalResult[i] = temp.get(i);
+            }
+            return finalResult;
+        }
+    }
+
+    public static long getUsersCount(Start start) throws SQLException {
+        String QUERY =
+                "SELECT COUNT(*) as total FROM " +
+                        Config.getConfig(start).getUsersTable();
+        try (Connection con = ConnectionPool.getConnection(start);
+             PreparedStatement pst = con.prepareStatement(QUERY)) {
+            ResultSet result = pst.executeQuery();
+            if (result.next()) {
+                return result.getLong("total");
+            }
+            return 0;
+        }
     }
 
     private static class PasswordResetTokenInfoRowMapper implements RowMapper<PasswordResetTokenInfo, ResultSet> {